requirements.txt
opencv-python==4.7.0.72
numpy==1.24.3
pyserial==3.5
mss==9.0.1
pyautogui==0.9.54
keyboard==0.13.5
pywin32==306
